cleanup and fixes
fix circular reference and other cleanup
